evals/acme_data_pipeline/app/requirements.txt
google-cloud-bigquery>=3.0
google-cloud-storage>=2.0
google-cloud-kms>=2.0
